src/train/requirements.txt
catboost
pandas
numpy
scikit-learn
python-dotenv
requests